backend/requirements.txt
asgiref==3.7.2
boto3==1.34.9
botocore==1.34.9
Django==4.2.7
django-environ==0.11.2
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Faker==21.0.0
jmespath==1.0.1
Pillow==10.1.0
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
s3transfer==0.10.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.0.7
